0.0.9.1b2 - now nu_install actually can _just_ install and doesn't reroll the distro unless needed
nu_install in b1 was really just nu_build. it always rerolled the
distro (even if it didn't always recompile). now nu_install can install
a finished build again and again, quite quickly. it still falls back to
building the distro if needed, and that still falls back to compiling
the source as needed, and that still falls back to fetching the source
and ports as needed.